+++ src/components/postDetailPage/BackToListBtn.tsx
@@ -1,8 +1,6 @@
-
 import { Post } from '@/types/post';
 import { Button } from '../ui/button';
 import { Link } from 'react-router-dom';
-
 
 type BackToListBtnProps = {
   data: Post;
