+++ src/hooks/queries/useAddBookmarkMutation.ts
@@ -2,13 +2,15 @@ import { addBookmark } from '@/apis/myPage/bookmark';
 // import { BookMark } from '@/types/bookmark';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-export const useAddBookmarkMutation = () => {
+export const useAddBookmarkMutation = (currentPostId: number | null) => {
   const queryClient = useQueryClient();
 
   return useMutation({
     mutationFn: (bookmarkId: number) => addBookmark(bookmarkId),
     onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['bookmarks'] });
+      if (currentPostId) {
+        queryClient.invalidateQueries({ queryKey: ['post', currentPostId] });
+      }
     },
     onError: () => {
       console.log('error!');
